Move "dry" communication tests to dedicated file
